refactor(i18n): move locales + related components to i18n folder
