+++ rspack-ui/src/i18n/locales.ts
