+++ rspack-ui/src/i18n/locales/en/messages.d.ts
