+++ rspack-ui/src/i18n/locales/en/messages.po
@@ -1,11 +1,11 @@
 msgid ""
 msgstr ""
-"POT-Creation-Date: 2023-08-23 14:57-0400\n"
+"POT-Creation-Date: 2023-08-24 12:53-0400\n"
 "MIME-Version: 1.0\n"
 "Content-Type: text/plain; charset=utf-8\n"
 "Content-Transfer-Encoding: 8bit\n"
 "X-Generator: @lingui/cli\n"
-"Language: fr\n"
+"Language: en\n"
 "Project-Id-Version: \n"
 "Report-Msgid-Bugs-To: \n"
 "PO-Revision-Date: \n"
@@ -15,16 +15,16 @@ msgstr ""
 
 #: src/Inbox.tsx:30
 msgid "{messagesCount, plural, one {There's # message in your inbox.} other {There are # messages in your inbox.}}"
-msgstr "{messagesCount, plural, one {Il y a # message dans boîte de réception.} other {Il y a # messages dans votre boîte de réception.}}"
+msgstr "{messagesCount, plural, one {There's # message in your inbox.} other {There are # messages in your inbox.}}"
 
 #: src/Inbox.tsx:32
 msgid "Last login on {0}."
-msgstr "Dernière connexion le {0}."
+msgstr "Last login on {0}."
 
 #: src/Inbox.tsx:21
 msgid "Message Inbox"
-msgstr "Boîte de réception de messages."
+msgstr "Message Inbox"
 
 #: src/Inbox.tsx:23
 msgid "See all <0>unread messages </0>or <1>mark them</1> as read."
-msgstr "Voir tous <0>les messages non lus </0>ou <1>les marquer</1> comme lus."
+msgstr "See all <0>unread messages </0>or <1>mark them</1> as read."
