+++ rspack-ui/src/i18n/locales/en/messages.ts
