+++ rspack-ui/src/i18n/locales/fr/messages.d.ts
