+++ rspack-ui/src/i18n/locales/fr/messages.ts
@@ -1,0 +1,1 @@
+/*eslint-disable*/export const messages=JSON.parse("{\"iHvPFN\":[[\"messagesCount\",\"plural\",{\"one\":[\"Il y a \",\"#\",\" message dans boîte de réception.\"],\"other\":[\"Il y a \",\"#\",\" messages dans votre boîte de réception.\"]}]],\"ItXLVU\":[\"Dernière connexion le \",[\"0\"],\".\"],\"8bWV5m\":\"Boîte de réception de messages.\",\"f8BUjV\":\"Voir tous <0>les messages non lus </0>ou <1>les marquer</1> comme lus.\"}");
