+++ rspack-ui/src/main.tsx
@@ -3,8 +3,8 @@ import { createRoot } from "react-dom/client";
 
 import { i18n } from "@lingui/core";
 import { I18nProvider } from "@lingui/react";
-import { messages as enMessages } from "./locales/en/messages";
-import { messages as frMessages } from "./locales/fr/messages";
+import { messages as enMessages } from "./i18n/locales/en/messages";
+import { messages as frMessages } from "./i18n/locales/fr/messages";
 
 import Inbox from "./Inbox";
